requirements.txt

# Django==5.1.4

mysqlclient==2.2.6

python-dotenv==0.21.0

pycryptodome==3.21.0

argon2-cffi==21.3.0

python-dateutil==2.8.2

requests==2.32.3

Django>=4.1.3,<4.2.0